Feature - show AA in project list
* updated proj list formatter to show legacy stup beneath title
* show if held elsewhere
* show AA ests
